# Optimize float to Decimal conversions