With this fix float/double to decimal conversion is faster than native code.
Contributes to https://github.com/dotnet/coreclr/issues/18249.

### 64-bit

|  FromFloat |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 19.09 ns | 0.1653 ns | 0.0590 ns |   1.00 |
|  CoreRT | 26.13 ns | 0.0675 ns | 0.0241 ns |   1.37 |
| CoreRT2 | 10.67 ns | 0.3193 ns | 0.1139 ns |   0.56 |

|  FromDouble |      Mean |     Error |    StdDev | Scaled |
|-------- |----------:|----------:|----------:|-------:|
|  Native | 63.310 ns | 0.1139 ns | 0.0406 ns |   1.00 |
|  CoreRT | 84.985 ns | 0.2272 ns | 0.0810 ns |   1.34 |
| CoreRT2 |  8.871 ns | 0.1378 ns | 0.0491 ns |   0.14 |

### 32-bit

|  FromFloat |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 22.92 ns | 0.2139 ns | 0.0763 ns |   1.00 |
|  CoreRT | 35.33 ns | 0.3903 ns | 0.1392 ns |   1.54 |
| CoreRT2 | 12.05 ns | 0.1421 ns | 0.0507 ns |   0.53 |

|  FromDouble |      Mean |     Error |    StdDev | Scaled |
|-------- |----------:|----------:|----------:|-------:|
|  Native |  38.48 ns | 0.1048 ns | 0.0162 ns |   1.00 |
|  CoreRT | 111.65 ns | 0.1547 ns | 0.0552 ns |   2.90 |
| CoreRT2 |  31.31 ns | 0.1573 ns | 0.0561 ns |   0.81 |