@AndyAyersMS 2018-06-02
There is something going on here that lead you to pull the intrinsic check out of the leading zeros helper. I believe I know the reason but would like to hear your thoughts on it first.