@pentp 2018-06-02
Crossgen didn't inline LeadingZeroCount but the JIT cost is still there, so it made more sense to force the entire function out of crossgen.