@jkotas 2018-06-02
Once we turn on tiered JIT by default, it is going to take care of cases like these (the crosssgened code is like "tier-0" code).