@tannergooding 2018-06-07
FYI. @CarolEidt, @eerhardt, This is probably the first dependency on the intrinsics.