@EgorBo 2018-06-07
@tannergooding so a typical intrinsics usage might look like:
```csharp
if (Arm.Simd.IsSupported)
    Arm.Simd..
else if (X86.SseX.IsSupported)
    X86.SseX..
else if (X86.Avx.IsSupported)
    X86.Avx..
else
    default impl
```
? 🙂