@tannergooding 2018-06-07
You would switch avx and ssex, but yes.