@AndyAyersMS 2018-06-07
We had some discussion over on gitter during the hackathon about the implications of using HW intrinsics in code that will be prejitted, like most code in corelib.

Wrapping HW intrinsics into helper methods (like the "before" version of `LeadingZeroCount`) means those helpers won't be inlinable in prejitted code,. The helper method callers will get prejitted. So despite this looking like a nice way to abstract over HW intrinsics it is not currently a very performant arrangement.

As Jan notes tiering can help by giving the system a chance to rejit the helper method callers, but we need to ensure that if we go down that route that the helper methods actually end up getting inlined. And we probably need some AOT friendly solution too.

In this change @pentp hoists the HW intrinsic uses up into the calling method. This ensures that method won't be prejitted and the intrinsics end up where we want them and we get good perf. This is fine in small doses. But I'm not sure we want this pattern to proliferate as we implement similar intrinsics for other platforms.

So, something to think about further....