@MichalStrehovsky 2018-06-07
@CarolEidt @tannergooding Would it be possible to prioritize dotnet/coreclr#15922 a bit higher so that we don't have to put `#if CORECLR` around code like this? Standing up the intrinsics in the CoreRT compiler with RyuJIT backend shouldn't be a big deal, and we can use the `IsSupported == false` implementation everywhere else. It would likely also help testability of changes like this. Changes under `#if CORECLR` in unshared files don't build in any configuration in this repo.