@eerhardt 2018-06-07
@maryamariyan - Is https://github.com/dotnet/coreclr/issues/15922 something you can take a look as part of your effort to move more source files into the shared folder?