@EgorBo 2018-07-09
```
(int)X86.Lzcnt.LeadingZeroCount(rgulRes[iHiRes]) : LeadingZeroCount(rgulRes[iHiRes]);
```
LeadingZeroCount(0) != X86.Lzcnt.LeadingZeroCount(0)
just in case...