@pentp 2018-07-10
`rgulRes[iHiRes]` is never zero, although it's hard to see that right now (an assert would help probably)